Move menu items to own components
